nodejs/oauthtest/requirements.txt
azure-identity
pyodbc
flask
msal
pytz
